django-app/src/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
pkg-resources==0.0.0
psycopg2-binary==2.7.4
pytz==2018.3
